pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hosthawk"
version = "1.0.0"
description = "Enterprise-grade network scanner for security professionals"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "HostHawk Team", email = "team@hosthawk.io"}
]
maintainers = [
    {name = "HostHawk Team", email = "team@hosthawk.io"}
]
keywords = ["network", "scanner", "security", "pentest", "vulnerability"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Networking",
    "Topic :: System :: Networking :: Monitoring",
]

dependencies = [
    "scapy>=2.5.0",
    "python-nmap>=0.7.1",
    "netifaces>=0.11.0",
    "python-whois>=0.8.0",
    "xmltodict>=0.13.0",
    "colorama>=0.4.6",
    "Flask>=2.0.1",
    "Flask-SocketIO>=5.1.1",
    "python-socketio>=5.4.0",
    "python-engineio>=4.2.1",
    "Werkzeug>=2.0.1",
    "Jinja2>=3.0.1",
    "MarkupSafe>=2.0.1",
    "pysnmp>=4.4.12",
    "pysnmp-mibs>=0.1.8",
    "pysmi>=0.3.4",
    "dpkt>=1.9.7.2",
    "pypcap>=1.2.3",
    "netaddr>=0.8.0",
    "requests>=2.26.0",
    "pandas>=1.3.3",
    "numpy>=1.21.2",
    "cryptography>=3.4.7",
    "pyOpenSSL>=20.0.1",
    "beautifulsoup4>=4.9.3",
    "dnspython>=2.1.0",
    "fpdf>=1.7.2",
]

[project.optional-dependencies]
dev = [
    "black>=21.7b0",
    "flake8>=3.9.2",
]

[project.urls]
Homepage = "https://github.com/yourusername/hosthawk"
Documentation = "https://hosthawk.readthedocs.io"
Repository = "https://github.com/yourusername/hosthawk"
"Bug Tracker" = "https://github.com/yourusername/hosthawk/issues"

[project.scripts]
hosthawk = "scanner.cli:main"

[tool.setuptools]
packages = ["scanner", "scanner.core", "scanner.plugins", "scanner.output", "scanner.reporting", "scanner.utils"]

[tool.setuptools.package-data]
scanner = ["reporting/templates/**/*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

